Breaks up LogViewer into multiple components
