+++ src/components/LogViewer/index.scss
@@ -257,22 +257,6 @@
       .value {
         color: #005ae5;
       }
-
-      //&.start, &.end {
-      //  color: rgba(0, 255, 0, 1);
-      //}
-      //
-      //&.object-start {
-      //  color: rgba(255, 0, 255, 1);
-      //}
-      //
-      //&.object-end {
-      //  color: rgba(255, 0, 0, 1);
-      //}
-      //
-      //&.literal {
-      //  color: rgba(0, 0, 255, 1);
-      //}
     }
   }
 }
